+++ js/main.js
@@ -1,7 +1,7 @@
 import { showPhotos } from './show-photos.js';
 import { uploadPhoto } from './upload-photo.js';
 import { fetchPhoto } from './fetch-photo.js';
-import { errorHandlerModal } from './error-handler.js';
+import { errorHandlerModal } from './notification/error-handler.js';
 
 fetchPhoto()
   .then(photos => {
